largest_histogram_area
take a stack and 2 vector.iterate the loop and check that untill find the smaller area on left keep going and save the indexes. untill the smaller area elment we get.the top +1 save on left array.

similar for right array..and iterate right to left.and save top-1;

and last (right[i]-left[i]-1)*height[i];